jetson/testrig/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "testrig"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio==1.80.0",
    "grpcio-tools==1.80.0",
    "protobuf==6.33.6",
    "pydantic==2.13.4",
    "psutil>=5.9.0",
]
